# backend/src/web/api/handlers/questions.rs
use axum::{
    extract::{Path, Query, State},
    http::StatusCode,
    response::IntoResponse,
    Json,
};
use serde::Deserialize;
use std::collections::HashMap;
use uuid::Uuid;

use crate::web::routes::AppState;
use crate::web::api::error::ApiError;
use crate::web::api::models::*;

#[derive(Debug, Deserialize)]
pub struct QuestionRevisionQuery {
    question_revision_id: Option<Uuid>,
}

/// List questions with latest revisions
#[utoipa::path(
    get,
    path = "/questions",
    tag = "Question",
    responses((status = 200, description = "Questions list", body = QuestionListResponse))
)]
pub async fn list_questions(
    State(app_state): State<AppState>,
) -> Result<Json<QuestionListResponse>, ApiError> {
    // Fetch all questions from the database
    let db_questions = app_state
        .database
        .questions
        .get_all_questions()
        .await
        .map_err(|e| ApiError::InternalServerError(format!("Failed to fetch questions: {e}")))?;

    // Convert database models to API models with their latest revisions
    let mut questions = Vec::new();
    for db_question in db_questions {
        // Fetch the latest revision for this question
        let revision_model = app_state
            .database
            .questions_revisions
            .get_latest_revision_by_question(db_question.question_id)
            .await
            .map_err(|e| {
                ApiError::InternalServerError(format!("Failed to fetch question revision: {e}"))
            })?;

        if let Some(revision_model) = revision_model {
            // Convert the JSON text to HashMap<String, String>
            let text_map = if let Some(text_obj) = revision_model.text.as_object() {
                text_obj
                    .iter()
                    .filter_map(|(k, v)| v.as_str().map(|s| (k.clone(), s.to_string())))
                    .collect()
            } else {
                HashMap::new()
            };

            let category = app_state
                .database
                .category_catalog
                .get_category_catalog_by_id(db_question.category_id)
                .await
                .map_err(|e| ApiError::InternalServerError(format!("Failed to fetch category: {e}")))?
                .ok_or(ApiError::InternalServerError("Category not found for question".to_string()))?;

            let question = Question {
                question_id: db_question.question_id,
                category: category.name,
                created_at: db_question.created_at.to_rfc3339(),
                latest_revision: QuestionRevision {
                    question_revision_id: revision_model.question_revision_id,
                    question_id: revision_model.question_id,
                    text: text_map,
                    weight: revision_model.weight as f64,
                    created_at: revision_model.created_at.to_rfc3339(),
                },
            };

            questions.push(question);
        }
    }

    Ok(Json(QuestionListResponse { questions }))
}

/// Create a question and initial revision
#[utoipa::path(
    post,
    path = "/questions",
    tag = "Question",
    request_body = CreateQuestionRequest,
    responses((status = 201, description = "Question created", body = QuestionResponse), (status = 400, description = "Validation error"))
)]
pub async fn create_question(
    State(app_state): State<AppState>,
    Json(request): Json<CreateQuestionRequest>,
) -> Result<impl IntoResponse, ApiError> {
    // Validate request
    if request.text.is_empty() {
        return Err(ApiError::BadRequest("Text must not be empty".to_string()));
    }

    if request.weight <= 0.0 {
        return Err(ApiError::BadRequest(
            "Weight must be greater than 0".to_string(),
        ));
    }

    // Create the question in the database
    let question_model = app_state
        .database
        .questions
        .create_question(request.category_id)
        .await
        .map_err(|e| ApiError::InternalServerError(format!("Failed to create question: {e}")))?;

    // Create the initial revision with the multilingual text and weight
    let text_json = serde_json::to_value(&request.text)
        .map_err(|e| ApiError::InternalServerError(format!("Failed to serialize text: {e}")))?;
    let revision_model = app_state
        .database
        .questions_revisions
        .create_question_revision(question_model.question_id, text_json, request.weight as f32)
        .await
        .map_err(|e| {
            ApiError::InternalServerError(format!("Failed to create question revision: {e}"))
        })?;

    // Build the response with the created question and revision
    let category = app_state
        .database
        .category_catalog
        .get_category_catalog_by_id(question_model.category_id)
        .await
        .map_err(|e| ApiError::InternalServerError(format!("Failed to fetch category: {e}")))?
        .ok_or(ApiError::InternalServerError("Category not found for question".to_string()))?;

    let question = Question {
        question_id: question_model.question_id,
        category: category.name,
        created_at: question_model.created_at.to_rfc3339(),
        latest_revision: QuestionRevision {
            question_revision_id: revision_model.question_revision_id,
            question_id: revision_model.question_id,
            text: request.text,
            weight: revision_model.weight as f64,
            created_at: revision_model.created_at.to_rfc3339(),
        },
    };

    Ok((StatusCode::CREATED, Json(QuestionResponse { question })))
}

/// Get a question by ID (optionally by revision)
#[utoipa::path(
    get,
    path = "/questions/{question_id}",
    tag = "Question",
    params(
        ("question_id" = uuid::Uuid, Path, description = "Question ID"),
        ("question_revision_id" = Option<uuid::Uuid>, Query, description = "Specific revision ID")
    ),
    responses((status = 200, description = "Question detail", body = QuestionResponse), (status = 404, description = "Not found"))
)]
pub async fn get_question(
    State(app_state): State<AppState>,
    Path(question_id): Path<Uuid>,
    Query(query): Query<QuestionRevisionQuery>,
) -> Result<Json<QuestionResponse>, ApiError> {
    // Fetch the question from the database
    let question_model = app_state
        .database
        .questions
        .get_question_by_id(question_id)
        .await
        .map_err(|e| ApiError::InternalServerError(format!("Failed to fetch question: {e}")))?;

    let question_model = match question_model {
        Some(q) => q,
        None => return Err(ApiError::NotFound("Question not found".to_string())),
    };

    // Fetch the revision for this question (specific revision if provided, otherwise latest)
    let revision = if let Some(revision_id) = query.question_revision_id {
        // Fetch specific revision by ID
        let revision = app_state
            .database
            .questions_revisions
            .get_revision_by_id(revision_id)
            .await
            .map_err(|e| {
                ApiError::InternalServerError(format!("Failed to fetch question revision: {e}"))
            })?;

        // Verify that the revision belongs to the requested question
        match revision {
            Some(r) if r.question_id == question_id => Some(r),
            Some(_) => {
                return Err(ApiError::BadRequest(
                    "Revision does not belong to the specified question".to_string(),
                ))
            }
            None => {
                return Err(ApiError::NotFound(
                    "Question revision not found".to_string(),
                ))
            }
        }
    } else {
        // Fetch latest revision for this question
        app_state
            .database
            .questions_revisions
            .get_latest_revision_by_question(question_id)
            .await
            .map_err(|e| {
                ApiError::InternalServerError(format!("Failed to fetch question revision: {e}"))
            })?
    };

    let revision = match revision {
        Some(r) => r,
        None => {
            return Err(ApiError::InternalServerError(
                "Question has no revisions".to_string(),
            ))
        }
    };

    // Convert the database text (JSON Value) to HashMap<String, String>
    let text_map: HashMap<String, String> = serde_json::from_value(revision.text).map_err(|e| {
        ApiError::InternalServerError(format!("Failed to parse question text: {e}"))
    })?;

    // Build the response
    let category = app_state
        .database
        .category_catalog
        .get_category_catalog_by_id(question_model.category_id)
        .await
        .map_err(|e| ApiError::InternalServerError(format!("Failed to fetch category: {e}")))?
        .ok_or(ApiError::InternalServerError("Category not found for question".to_string()))?;

    let question = Question {
        question_id: question_model.question_id,
        category: category.name,
        created_at: question_model.created_at.to_rfc3339(),
        latest_revision: QuestionRevision {
            question_revision_id: revision.question_revision_id,
            question_id: revision.question_id,
            text: text_map,
            weight: revision.weight as f64,
            created_at: revision.created_at.to_rfc3339(),
        },
    };

    Ok(Json(QuestionResponse { question }))
}

/// Update a question (creates a new revision)
#[utoipa::path(
    put,
    path = "/questions/{question_id}",
    tag = "Question",
    params(("question_id" = uuid::Uuid, Path, description = "Question ID")),
    request_body = UpdateQuestionRequest,
    responses((status = 200, description = "Question updated", body = QuestionResponse), (status = 404, description = "Not found"), (status = 400, description = "Validation error"))
)]
pub async fn update_question(
    State(app_state): State<AppState>,
    Path(question_id): Path<Uuid>,
    Json(request): Json<UpdateQuestionRequest>,
) -> Result<Json<QuestionResponse>, ApiError> {
    // Validate request
    if request.text.is_empty() {
        return Err(ApiError::BadRequest("Text must not be empty".to_string()));
    }

    if request.weight <= 0.0 {
        return Err(ApiError::BadRequest(
            "Weight must be greater than 0".to_string(),
        ));
    }

    // Update the question category in the database
    let updated_question_model = app_state
        .database
        .questions
        .update_question(question_id, Some(request.category_id))
        .await
        .map_err(|e| {
            if e.to_string().contains("Question not found") {
                ApiError::NotFound("Question not found".to_string())
            } else {
                ApiError::InternalServerError(format!("Failed to update question: {e}"))
            }
        })?;

    // Create a new revision for the question update
    let text_json = serde_json::to_value(&request.text)
        .map_err(|e| ApiError::InternalServerError(format!("Failed to serialize text: {e}")))?;
    let revision_model = app_state
        .database
        .questions_revisions
        .create_question_revision(question_id, text_json, request.weight as f32)
        .await
        .map_err(|e| {
            ApiError::InternalServerError(format!("Failed to create question revision: {e}"))
        })?;

    // Build the response
    let category = app_state
        .database
        .category_catalog
        .get_category_catalog_by_id(updated_question_model.category_id)
        .await
        .map_err(|e| ApiError::InternalServerError(format!("Failed to fetch category: {e}")))?
        .ok_or(ApiError::InternalServerError("Category not found for question".to_string()))?;

    let question = Question {
        question_id: updated_question_model.question_id,
        category: category.name,
        created_at: updated_question_model.created_at.to_rfc3339(),
        latest_revision: QuestionRevision {
            question_revision_id: revision_model.question_revision_id,
            question_id: revision_model.question_id,
            text: request.text,
            weight: revision_model.weight as f64,
            created_at: revision_model.created_at.to_rfc3339(),
        },
    };

    Ok(Json(QuestionResponse { question }))
}

/// Delete a question revision by ID
#[utoipa::path(
    delete,
    path = "/questions/revisions/{revision_id}",
    tag = "Question",
    params(("revision_id" = uuid::Uuid, Path, description = "Question revision ID")),
    responses((status = 204, description = "Deleted"), (status = 400, description = "In use"), (status = 404, description = "Not found"))
)]
pub async fn delete_question_revision_by_id(
    State(app_state): State<AppState>,
    Path(revision_id): Path<Uuid>,
) -> Result<StatusCode, ApiError> {
    // Check if the question revision exists
    let revision_exists = app_state
        .database
        .questions_revisions
        .get_revision_by_id(revision_id)
        .await
        .map_err(|e| ApiError::InternalServerError(format!("Failed to fetch question revision: {e}")))?;

    if revision_exists.is_none() {
        return Err(ApiError::NotFound("Question revision not found".to_string()));
    }

    // Check if any assessment responses reference this question revision
    let has_responses = app_state
        .database
        .assessments_response
        .has_responses_for_question_revision(revision_id)
        .await
        .map_err(|e| ApiError::InternalServerError(format!("Failed to check assessment responses: {e}")))?;

    if has_responses {
        return Err(ApiError::BadRequest(
            "Cannot delete question revision: It is currently being used in assessment responses. Please ensure no active assessments reference this revision before deletion.".to_string()
        ));
    }

    // Delete the question revision from the database
    app_state
        .database
        .questions_revisions
        .delete_revision(revision_id)
        .await
        .map_err(|e| ApiError::InternalServerError(format!("Failed to delete question revision: {e}")))?;

    Ok(StatusCode::NO_CONTENT)

}