# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Assessment System API",
    "description": "A comprehensive API for managing assessments, submissions, and reviews in a sustainability assessment system.\n\n## Authentication\nAll endpoints require JWT authentication. Include the token in the Authorization header:\n`Authorization: Bearer <your-jwt-token>`\n\n## Versioning\nThe API supports response versioning to track changes and resolve conflicts.\n",
    "contact": {
      "name": "Assessment System Team",
      "email": "support@assessment-system.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/questions": {
      "get": {
        "tags": ["Questions"],
        "summary": "List questions",
        "description": "Get all questions",
        "responses": {
          "200": {
            "description": "List of questions with their revisions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListWithRevisionsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Questions"],
        "summary": "Create question",
        "description": "Create a new question (Admin only)",
        "operationId": "create_question",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Question created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/questions/{question_id}": {
      "get": {
        "tags": ["Questions"],
        "summary": "Get question",
        "description": "Get specific question with all revisions or a specific revision",
        "operationId": "get_question",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "description": "Question ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "revision_id",
            "in": "query",
            "description": "Specific revision ID to retrieve (optional)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "Language for content",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionWithRevisionsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": ["Questions"],
        "summary": "Update question",
        "description": "Update a question (Admin only). Include question_revision_id for optimistic locking.",
        "operationId": "update_question",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "description": "Question ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "question_revision_id",
            "in": "query",
            "description": "Current revision ID for optimistic locking",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Question updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Conflict - revision mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/questions/revisions/{question_revision_id}": {
      "delete": {
        "tags": ["Questions"],
        "summary": "Delete question revision",
        "description": "Delete a question revision (Admin only). Requires question_revision_id as path parameter.",
        "operationId": "delete_question_revision",
        "parameters": [
          {
            "name": "question_revision_id",
            "in": "path",
            "description": "Question Revision ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Question revision deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Conflict - revision mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Question": {
        "required": ["question_id", "category_id", "created_at", "latest_revision"],
        "type": "object",
        "properties": {
          "question_id": {
            "type": "string",
            "format": "uuid"
          },
          "category_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "latest_revision": {
            "$ref": "#/components/schemas/QuestionRevision"
          }
        }
      },
      "QuestionRevision": {
        "required": [
          "question_revision_id",
          "question_id",
          "text",
          "weight",
          "created_at"
        ],
        "type": "object",
        "properties": {
          "question_revision_id": {
            "type": "string",
            "format": "uuid"
          },
          "question_id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "object"
          },
          "weight": {
            "type": "number"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateQuestionRequest": {
        "required": ["category_id", "text"],
        "type": "object",
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "object",
            "default": {
              "en": "explain your ecological policies"
            }
          },
          "weight": {
            "type": "number",
            "default": 1.0
          }
        }
      },
      "UpdateQuestionRequest": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "object"
          },
          "weight": {
            "type": "number"
          }
        }
      },
...
      "QuestionListWithRevisionsResponse": {
        "required": ["questions"],
        "type": "object",
        "properties": {
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionWithRevisionsResponse"
            }
          }
        }
      },
      "QuestionResponse": {
        "required": ["question"],
        "type": "object",
        "properties": {
          "question": {
            "$ref": "#/components/schemas/Question"
          }
        }
      },
      "QuestionWithRevisionsResponse": {
        "required": ["question", "revisions"],
        "type": "object",
        "properties": {
          "question": {
            "$ref": "#/components/schemas/Question"
          },
          "revisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionRevision"
            }
          }
        }
      },
...
      "Error": {
        "required": ["error", "message"],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true
          }
        }
      },
      "ValidationError": {
        "required": ["error", "message", "validation_errors"],
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "validation_failed"
          },
          "message": {
            "type": "string",
            "example": "Request validation failed"
          },
          "validation_errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError_validation_errors"
            }
          }
        }
      },
...
      "ValidationError_validation_errors": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}